EventRush/Python/src/requirements.txt
Pillow
glob2
numpy